requirements.txt
django==4.2
djangorestframework==3.15.2
python==3.12.5